Ant based solution for mr-jars.
