--- jaxb-ri/jaxb-xjc-jdk9/src/test/resources/com/sun/tools/xjc/resources/xjc.xml
@@ -1,100 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
-
-    Copyright (c) 2017 Oracle and/or its affiliates. All rights reserved.
-
-    The contents of this file are subject to the terms of either the GNU
-    General Public License Version 2 only ("GPL") or the Common Development
-    and Distribution License("CDDL") (collectively, the "License").  You
-    may not use this file except in compliance with the License.  You can
-    obtain a copy of the License at
-    https://oss.oracle.com/licenses/CDDL+GPL-1.1
-    or LICENSE.txt.  See the License for the specific
-    language governing permissions and limitations under the License.
-
-    When distributing the software, include this License Header Notice in each
-    file and include the License file at LICENSE.txt.
-
-    GPL Classpath Exception:
-    Oracle designates this particular file as subject to the "Classpath"
-    exception as provided by Oracle in the GPL Version 2 section of the License
-    file that accompanied this code.
-
-    Modifications:
-    If applicable, add the following below the License Header, with the fields
-    enclosed by brackets [] replaced by your own identifying information:
-    "Portions Copyright [year] [name of copyright owner]"
-
-    Contributor(s):
-    If you wish your version of this file to be governed by only the CDDL or
-    only the GPL Version 2, indicate your decision by adding "[Contributor]
-    elects to include this software in this distribution under the [CDDL or GPL
-    Version 2] license."  If you don't indicate a single choice of license, a
-    recipient has the option to distribute your version of this file under
-    either the CDDL, the GPL Version 2 or to extend the choice of license to
-    its licensees as provided above.  However, if you add GPL Version 2 code
-    and therefore, elected the GPL Version 2 license, then the option applies
-    only if the new code is made subject to such option by the copyright
-    holder.
-
--->
-
-<project name="xjctest" default="test" basedir=".">
-
-    <target name="test">
-        <echoproperties/>
-    </target>
-
-    <target name="init">
-        <property name="libs.folder" value="${bin.folder}/lib"/>
-        <property name="endorsed.folder" value="${bin.folder}/endorsed"/>
-        <property name="lib.dir" value="lib"/>
-        <property name="build.dir" value="build"/>
-        <property name="src.dir" value="src"/>
-        <property name="schema" value="simple.xsd"/>
-    </target>
-
-    <target name="setup" depends="init">
-        <mkdir dir="${lib.dir}"/>
-        <mkdir dir="${build.dir}"/>
-        <mkdir dir="${src.dir}"/>
-        <copy todir="${lib.dir}">
-            <fileset dir="${libs.folder}"/>
-        </copy>
-        <path id="task.cp">
-            <pathelement location="${bin.folder}/classes"/>
-            <fileset dir="${lib.dir}"/>
-            <fileset dir="${endorsed.folder}"/>
-        </path>
-    </target>
-
-    <target name="xjc-init" depends="setup">
-        <taskdef name="xjc" classname="com.sun.tools.xjc.XJCTask">
-            <classpath refid="task.cp"/>
-        </taskdef>
-    </target>
-
-    <target name="xjc-fork" depends="xjc-init">
-        <echo message="Compiling the schema..."/>
-        <xjc schema="${schema}" package="simple" destdir="${build.dir}" fork="true">
-            <produces dir="${build.dir}/simple" includes="**/*.java"/>
-            <jvmarg line="--add-modules java.activation"/>
-        </xjc>
-    </target>
-
-    <target name="xjc-addmodules" depends="xjc-init">
-        <echo message="Compiling the schema..."/>
-        <xjc schema="${schema}" package="simple" destdir="${build.dir}" fork="true" addmodules="java.activation">
-            <produces dir="${build.dir}/simple" includes="**/*.java"/>
-        </xjc>
-    </target>
-
-    <target name="clean">
-        <delete dir="${src.dir}" verbose="true" failonerror="false"/>
-        <delete dir="${build.dir}" verbose="true" failonerror="false"/>
-        <delete dir="${lib.dir}" verbose="true" failonerror="false"/>
-    </target>
-
-</project>
